移除uniform量 加入dataActor 加入pause 加入exit 加入in、out、incopy 性能优化 链接函数优化

+++ DMSL/DMSLUtils.h
@@ -20,8 +20,8 @@ namespace Dmsl {
 		//去除注释
 		std::string DeleteComments(const std::string&);
 
-		//转换到float
-		float ToFloat(const std::string&);
+		//转换到double
+		double ToDouble(const std::string&);
 
 		//转换到int
 		int ToInt(const std::string&);
